Added functionality to Navbar links: 'Dashboard' links to the dashboard page, and 'New Ticket' opens up TicketModal on both Dashboard page and individal ViewTicket pages. This means users can submit a ticket from any page on the site. Additionally, the 'Create Ticket' button that is visible on the Dashboard was moved from the TicketModal componenet (where it had previously been) to being on the Dashboard itself.
